@@ -15,10 +15,10 @@ package vinyldns
 import (
 	"crypto/tls"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"os"
 )
 
 type testToolsConfig struct {
@@ -28,7 +28,7 @@ type testToolsConfig struct {
 }
 
 func readFile(file string) (string, error) {
-	b, err := ioutil.ReadFile(file)
+	b, err := os.ReadFile(file)
 	if err != nil {
 		return "", err
 	}
