@@ -227,7 +227,7 @@ func TestRecordSetEP(t *testing.T) {
 }
 
 func TestRecordSetChangesEP(t *testing.T) {
-	rsc := recordSetChangesEP(c, "123", ListFilter{})
+	rsc := recordSetChangesEP(c, "123", ListFilterInt{})
 	expected := "http://host.com/zones/123/recordsetchanges"
 
 	if rsc != expected {
@@ -238,9 +238,9 @@ func TestRecordSetChangesEP(t *testing.T) {
 }
 
 func TestRecordSetChangesEPWithQuery(t *testing.T) {
-	rsc := recordSetChangesEP(c, "123", ListFilter{
+	rsc := recordSetChangesEP(c, "123", ListFilterInt{
 		MaxItems:  3,
-		StartFrom: "1",
+		StartFrom: 1,
 	})
 	expected := "http://host.com/zones/123/recordsetchanges?startFrom=1&maxItems=3"
 
