@@ -209,7 +209,7 @@ func (c *Client) RecordSetDelete(zoneID, recordSetID string) (*RecordSetUpdateRe
 }
 
 // RecordSetChanges retrieves the RecordSetChanges response for the Zone and ListFilter it's passed.
-func (c *Client) RecordSetChanges(zoneID string, f ListFilter) (*RecordSetChanges, error) {
+func (c *Client) RecordSetChanges(zoneID string, f ListFilterInt) (*RecordSetChanges, error) {
 	rsc := &RecordSetChanges{}
 	err := resourceRequest(c, recordSetChangesEP(c, zoneID, f), "GET", nil, rsc)
 	if err != nil {
@@ -221,7 +221,7 @@ func (c *Client) RecordSetChanges(zoneID string, f ListFilter) (*RecordSetChange
 
 // RecordSetChangesListAll retrieves the complete list of record set changes for the Zone ListFilter criteria passed.
 // Handles paging through results on the user's behalf.
-func (c *Client) RecordSetChangesListAll(zoneID string, filter ListFilter) ([]RecordSetChange, error) {
+func (c *Client) RecordSetChangesListAll(zoneID string, filter ListFilterInt) ([]RecordSetChange, error) {
 	if filter.MaxItems > 100 {
 		return nil, fmt.Errorf("MaxItems must be between 1 and 100")
 	}
@@ -237,7 +237,7 @@ func (c *Client) RecordSetChangesListAll(zoneID string, filter ListFilter) ([]Re
 		rsc = append(rsc, resp.RecordSetChanges...)
 		filter.StartFrom = resp.NextID
 
-		if len(filter.StartFrom) == 0 {
+		if filter.StartFrom == 0 {
 			return rsc, nil
 		}
 	}
